Move ReportForm into subfolder
